requirements.txt
discord.py==2.0.0
aiohttp==3.8.1
pydash==5.1.0
python-dotenv==0.21.0
deep_translator
motor